Personal quests > Quest approve complition (#17)

+++ src/QueReal.DAL/Models/QuestItem.cs
@@ -4,7 +4,7 @@
 	{
 		public string Title { get; set; }
 
-		public short Progress { get; set; } = 0;
+		public byte Progress { get; set; } = 0;
 
 		public Guid QuestId { get; set; }
 		public Quest Quest { get; set; }
